requirements.txt
asgiref==3.11.0
asttokens==3.0.1
decorator==5.2.1
Django==4.2.16
django-extensions==3.2.1
exceptiongroup==1.3.1
executing==2.2.1
factory-boy==3.2.0
Faker==8.10.1
ipython==8.29.0
jedi==0.19.2
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
parso==0.8.5
pexpect==4.9.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
text-unidecode==1.3
traitlets==5.14.3
typing_extensions==4.15.0
wcwidth==0.2.14
Werkzeug==3.1.3
